cli: Move ClientExport to cliutil
